+++ src/pydbm/models/fields/auto.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import hashlib
-import logging
 import typing
 
 from pydbm.models.fields.base import BaseField
@@ -9,10 +8,11 @@ from pydbm.models.fields.base import BaseField
 if typing.TYPE_CHECKING:
     from pydbm.models.base import BaseModel
     from pydbm.models.meta import Meta
+    from pydbm.typing_extra import SupportedClassT
 
-__all__ = ("AutoField",)
-
-logger = logging.getLogger(__name__)
+__all__ = (
+    "AutoField",
+)
 
 Self = typing.TypeVar("Self", bound="AutoField")  # unexport: not-public
 
@@ -23,11 +23,9 @@ class AutoField(BaseField):
         "fields",
     )
 
-    def __init__(
-        self, field_name: str, field_type_name: str, *, unique_together: tuple[str, ...] | None = None, **kwargs
-    ) -> None:
+    def __init__(self, field_name: str, field_type: SupportedClassT, *, unique_together: tuple[str, ...] | None = None, **kwargs) -> None:  # noqa: E501
         self.field_name = field_name
-        self.field_type_name = field_type_name
+        self.field_type = field_type
 
         self.public_name = field_name
         self.private_name = "_" + field_name
@@ -46,10 +44,10 @@ class AutoField(BaseField):
             raise ValueError("unique_together ise set, fields must be passed")
 
         self.fields = fields
-        return super().__call__(self.field_name, self.field_type_name, *args, **kwargs)  # type: ignore
+        return super().__call__(self.field_name, self.field_type, *args, **kwargs)  # type: ignore
 
     def generate_pk(self) -> str:
-        if self.unique_together:
+        if self.unique_together and self._is_call_run:
             text = "*".join(map(str, (attr for name in self.unique_together if (attr := self.fields.get(name, None)))))
             return hashlib.md5(bytes(text, "utf-8")).hexdigest()
         else: