+++ tests/models/fields/test_auto.py
@@ -1,17 +1,17 @@
 import pytest
 
-from pydbm import AutoField
+from pydbm.models.fields import AutoField
 
 
 def test_auto_field_repr():
-    field = AutoField("pk", "str")
+    field = AutoField("pk", str)
     assert repr(field) == "AutoField(default=Undefined, default_factory=generate_pk, normalizers=[], validators=[])"
-    assert repr(field()) == "AutoField('pk', 'str')"
+    assert repr(field()) == f"AutoField('pk', {str!r})"
 
 
 def test_auto_field():
     class Model:
-        pk = AutoField("pk", "str")()
+        pk = AutoField("pk", str)()
 
     model = Model()
     assert model.pk
@@ -21,7 +21,7 @@ def test_auto_field():
     assert str(cm.value) == "AutoField is read-only"
 
     class Model:
-        pk = AutoField("pk", "str", unique_together=("name", "username"))({"name": "hakan", "username": "hakancelik"})
+        pk = AutoField("pk", str, unique_together=("name", "username"))({"name": "hakan", "username": "hakancelik"})
 
     model = Model()
     assert model.pk == "e2cd7febd091a509284e8ab0fc302fbb"